+++ scripts/build-full.sh
@@ -12,8 +12,7 @@
 
 export NODE_PATH='./src'
 
-# export CMS_API='https://joplin-staging.herokuapp.com/api/graphql'
-export CMS_API='http://joplin-pr-3814-events-non-joyo.herokuapp.com/api/graphql'
+export CMS_API='https://joplin-staging.herokuapp.com/api/graphql'
 export CMS_MEDIA='https://joplin-austin-gov-static.s3.amazonaws.com/staging/media'
 
 yarn npm-run-all build-css build-js
